apps/stockpilot/requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1
aiosqlite==0.19.0
alembic==1.12.1

# AI 및 벡터 검색
openai==1.6.1
tiktoken==0.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# 데이터 처리 및 분석
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
yfinance==0.2.22
ta==0.10.2
plotly==5.17.0
matplotlib==3.8.2

# API 및 데이터 수집
requests==2.31.0
aiohttp==3.9.0
beautifulsoup4==4.12.2
python-multipart==0.0.6

# 보안 및 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# 작업 큐 및 스케줄링
celery==5.3.4
redis==5.0.1

# 로깅 및 모니터링
loguru==0.7.2
prometheus-client==0.19.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 환경 변수 관리
python-dotenv==1.0.0

# CORS 처리
fastapi-cors==0.0.6

# 시간대 처리
pytz==2023.3
python-dateutil==2.8.2

# ML 모델 저장/로드
joblib==1.3.2
pickle-mixin==1.0.2

# 암호화폐 데이터 (선택사항)
ccxt==4.1.52

# 뉴스 데이터 수집
newspaper3k==0.2.8
feedparser==6.0.10

# 이메일 알림
smtplib==3.6.0
email-validator==2.1.0